[PTI-SDK] Fix disabling XPTI when sycl runtime tracing is not enabled (#36)
Until the user enables sycl runtime tracing, XPTI should be disabled after it
is able to capture all the proper tracing streams (e.g. `sycl` and `sycl.pi`).

This fixes a bug where this operation was happening in the wrong place.
The `sycl.pi` stream does not have `graph_create` trace type. Instead, we moved
it to the `funcion_begin` trace type. This will happen when the first sycl
function is called.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>