[PTI-LIB] Fix FindLevelZero.cmake Version Parsing (#125)
As of more recent versions of Level Zero Loader, the `level-zero.pc`
file was removed from the installation. This fix uses the
(existing) `libze_loader.pc` file to parse the version instead.

Also, this includes a fix for a leak in the tests along with a test
improvement (calling EXPECT before ASSERT). Includes a change to the
configuration for picking up the correct `llvm-symbolizer`.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>