[PTI-LIB] Allow CMake to find LTS L0 Loader. (#170)
Allows CMake to find the Long Term Support version of the Level Zero
Loader. Ideally, this should not be used. However, it is an IPEX
requirement to link against the LTS version of the drivers.

There is some support via full tracing mode for this case but otherwise,
prefer using the rolling release version of the loader.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>