Embed OpenCL tracing files in the repo
Formerly these were obtained from a fixed commit from
https://github.com/intel/compute-runtime.git. This has no advantage
compared with directly embedding the files. It makes the tools hard to
build on a cluster without internet access.

The git hash from which the embedded copies are taken was
ca7d47598a63959d42750c62a2981e08ffa392a1 and has been recorded in
those files. That was the only modification to the files.

The copies were embedded in a subdirectory matching those from the
repo from which they were taken, so no modifications to these files
are needed in order to use them. All tools in this repo already add
the utils subdirectory as a directory in which the compiler should
search.

Also minor related changes and simplifications.