[PTI-LIB] Fix Python Commands in CMake/CTest (#288)
Fix logic for running the CTest when the Python interpreter was found.
The interpreter is REQUIRED by the project, so the conditional is
redundant. Also, the `Python_FOUND` variable refers to the entire Python
package (including development libraries). If this functionality is
desired in the future, the `Python_Interpreter_FOUND` variable is
better suited for this task.

Add logic to prevent multiple calls to `FindPython`.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>