Initial implementation of core hw metric collection interface (#211)
Simple core interface supports one metric group and dumps the processed data in a buffer. User can then be use the buffer create a json file that can be viewed in perfetto or any other format of interest. The interface currently supports only one metric group at a time and works with both time based streamer metric groups and trace based tracer metric groups. For the most part, it reuses the unitrace implementation for the time based metrics core functionality. The trace based metrics implementation follows the same flow.

The interface includes the following functions, some of whom are placeholders and not implemented yet
1- ptiMetricsGetDevices: Discovers available supported devices on the system
2- ptiMetricsGetMetricGroups: Gets all metric groups that can be collected on the selected supported device
3- ptiMetricsGetMetricsProperties: Gets properties for all metrics in a metric group
4- ptiMetricsConfigureCollection: Configures for collection the metric groups of interest on the selected supported device
5- ptiMetricsStartCollection: Starts the metrics collection on the selected supported device
6- ptiMetricsStartCollectionPaused: (Not implemented yet) Starts the metrics collection on the selected supported device in paused mode
7- ptiMetricsPauseCollection: (Not implemented yet) Pauses a started metrics collection on the selected supported device
8- ptiMetricsResumeCollection: (Not implemented yet) Resumes a paused metrics collection on the selected supported device
9- ptiMetricsStopCollection: Stops a started metrics collection on the selected supported device
10- ptiMetricsGetCalculatedData: Dumps processed data in a buffer after a collection is stopped on the selected device

Testing done: a modified version of iso3dfd_dpcpp is used to exercise the metric collection interface and to log the buffered data to a json file can be viewed in perfetto. The library implementation also generates a json file before buffering for comparison.