Fix read raw metrics buffer function (#379)
L0 ZeStreamer and ZeTracer Read functions are not able to capture how much data can be read from hw buffer without actually reading the data. Trying to capture the required size can results in data loss. This is the main reason the metrics_iso3dfd sample is failing on windows. Data was likely lost on Linux as well.
Instead of trying to capture the data size to allocate a local buffer of the correct size, tools need to just provide a large enough buffer and capture what can fit in the buffer. L0 Read functions use the in/out size variable to read up to the specified size and to notify how much data was read.

Same fix was also needed in unitrace

Testing done: Metric library and unitrace tested to work as expected
metrics_iso3dfd sample is left disable on windows as there still seems to be an issue when notifyEveryNReport is too big