Capture metrics without a ready event before closing the streamer/tracer (#400)
When reading the raw data from the hw buffer right before closing
the streamer or the tracer, we should not check the event for data
readiness. Instead, we should just capture whatever is in the buffer
even if the event buffer full is not signaled. Otherwise, data will be
lost right before closing the streamer/tracer.
This also affects unitrace and fixed in this patch as well.
Also enabled metrics iso3dfd on windows as part of this patch.