[PTI-LIB] Fix Assertion Executing Command Queue (#423)
Fix assertion hit while running IPEX on BMG.

When tracing starts after the command list is created and before the
command list gets executed on the command queue, an assertion is hit
because there is no command list stored in the `command_list_map_`. If
introspection is available, we can fix this.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>