[PTI-LIB] Don't run TSan on icpx Pre-2025.1 (#533)
Prevent ThreadSanitizer tests from running on nightly build and test of
PTI. Before, we were silencing the exit code of ThreadSanitizer via
`exitcode=0`. However, now we are trying to increase the visibility of
TSan errors.

This commit prevents a crash in 2025.0 and earlier ICPX compilers due to
lack of TSan support for unloading ignored libraries. This appears to
have been fixed here:
https://github.com/intel/llvm/commit/79c4ece75f5fcc8a9e6bcd560f3184cfab04378a

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>