[PTI-264] Fix kernel timestamp queries when counter based signal even… (#620)
* [PTI-264] Fix kernel timestamp queries when counter based signal event is used.

The same counter based event can be used multiple times as a signal event
within an in-order command list. However only the last signal of the event
can be waited on. So if such event is allocated from a kernel timestamp capable
event pool, quering the kernel timestamp retreives the timing only for the last
command within the command list that used the event as a signaling event.

In order to get timings for all command also in the above scenario, we do
the following:
1) Do not override event pool flags for counter based event pools
   (meaning, we do not force it to have kernel timestamps).
2) When a counter based event is used as the signal event for
   an appended command, we replace it with an event from the
   event cache, and appending a zeAppendSignalEvent command following
   the application appended command to signal the counter based event.
   That works since counter based events can be used only within
   in-order command lists.

Another small fixes:
1) Fixed seg-fault during API tracing when application calls
   zeKernelSetArgumentValue with pArgValue==NULL.
2) Commented out assert in ze_event_cache.h to workarround what
   seems to be a driver issue with immediate mode command lists
   and event reset from host.

* [PTI-264] Added a test to verify the scenario fixed in this PR.

This adds a unitrace test which check that the reported device
timeline, in --chrome-device-timeline scenario, has no execution overlaps
for counter-based, in order command list submissions.

Added two optional arguments to test/run_test.py script:
   usage: run_test.py [-h] [-s SCENARIO] [-t EXTRA_TEST_PROG] cmake_root_path args [args ...]

   -s SCENARIO - when specified the test will run only for the specified
                 scenario - other scenarios will get skipped (test will silently pass)
   -t EXTRA_TEST_PROG - specify additional program to invoke to verify the content of
                        unitrace generated files. The list of all output files are passed
                        as arguments to the command, like:
                          <EXTRA_TEST_PROG> file0 [file1 ...]

---------

Co-authored-by: Guy Zadicario <gzadicario@habana.ai>