[PTI-LIB] Increase stability of multithreaded test (#680)
Fixes a race condition and ThreadSanitizer failure in
`MultiThreadedSubmissionFixtureTest.MultiThreadedSubmissionTestUsingSyclOpsAllClassAllGroups`
on the 2025.3 oneAPI compiler.

Also, it reduces the flakiness of this test.

Moves `Local` mode handling before moving GPU command to list. This
prevents a race condition between removing the event from the list and handling
 local mode changes. Now the local mode changes are made before adding
to the list. This avoids needing to add another lock too.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>