[PTI-LIB] Fix Hang on BMG with SYCL 2025.3 URZ V2 (#691)
Fix hang on some tests with the Intel(R) oneAPI Compiler Runtime 2025.3.

This changes the behavior of Local mode to inject a signal to fix
this hang. Use `PTI_FORCE_BRIDGE_COMMANDS=1` to revert back to the old
beheavior.

Recently, the SYCL runtime switched to a different Level Zero backend
for SYCL (V2 backend). This backend is the default on BMG from 2025.3
release onwards.

The primary difference between V1 and V2 backends is the usage of Level Zero
[Counter Based Events](https://github.com/intel/compute-runtime/blob/master/level_zero/doc/experimental_extensions/COUNTER_BASED_EVENTS.md)
extension. With this new extension, our tests hang and we found that an
event is not being signaled.

Ideally, this would be only done for in_order Level Zero command lists.
We do not have the capability of querying for this
information. Right now, PTI Local mode only supports in_order command lists.
This change continues this trend and could make it even worse. Try
reverting to the old behavior if you would like to support ooo lists.

Signed-off-by: Schilling, Matthew <matthew.schilling@intel.com>